Refactored pdf
